app/requirements.txt
aiofiles
aioredis
aiosmtplib
anyio
asgiref
async-timeout
autopep8
bcrypt
blinker
certifi
cffi
charset-normalizer
click
colorama
cryptography
Deprecated
dnspython
ecdsa
email-validator
fakeredis
fastapi
fastapi-mail
greenlet
h11
httpcore
httptools
httpx
idna
Jinja2
MarkupSafe
packaging
passlib
pyasn1
pycodestyle
pycparser
pydantic
PyJWT
PyMySQL
pyparsing
python-dotenv
python-jose
python-multipart
PyYAML
redis
rfc3986
rsa
six
sniffio
sortedcontainers
SQLAlchemy
starlette
toml
typing_extensions
urllib3
uvicorn
watchgod
websockets
wrapt
